// packages/trpc/server/template-router/set-template-attachments.types.ts
import { z } from 'zod';

import { AttachmentType } from '@documenso/prisma/generated/types';

export const ZSetTemplateAttachmentsSchema = z.object({
  templateId: z.number(),
  attachments: z.array(
    z.object({
      id: z.string(),
      label: z.string().min(1, 'Label is required'),
      url: z.string().url('Invalid URL'),
      type: z.nativeEnum(AttachmentType),
    }),
  ),
});

export type TSetTemplateAttachmentsSchema = z.infer<typeof ZSetTemplateAttachmentsSchema>;

export const ZSetTemplateAttachmentsResponseSchema = z.array(
  z.object({
    id: z.string(),
    label: z.string(),
    url: z.string(),
    type: z.nativeEnum(AttachmentType),
  }),
);